pyproject.toml
[project]
name = "Genesis-Envs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "genesis-world>=0.2.1",
    "omegaconf>=2.3.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "ruff>=0.11.6",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tyro>=0.9.19",
]

[tool.setuptools.packages.find]
where = ["."]
